Update README.md
### Short Description for Tic Tac Toe Game

This Tic Tac Toe game is a two-player web application built with HTML, CSS and Javascipt. Players take turns placing X and O on a 3x3 grid, with the game detecting winning combinations and highlighting them. A restart button allows for quick restarts, and the responsive design ensures a smooth experience on all devices. It’s a fun, interactive way to engage in classic gameplay while serving as a great project for those learning web development.